backend/requirements.txt
langchain==0.0.183
langchain-community==0.0.4
langchain-groq==0.1.0
supabase==1.0.0
pandas==1.5.3
python-docx==0.8.11
python-pptx==0.6.21
torch==2.1.0
faiss-cpu==1.7.4
huggingface-hub==0.13.4
